# frontend/src/api/swagger-spec.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /api/check-in/get-current-parking:
    post:
      operationId: getCurrentParking
      summary: ''
      description: 現在地のパーキングの情報を取得する
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCurrentParkingRequest'
      responses:
        '200':
          description: 現在地のパーキングの情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parkings'
      tags: &ref_0
        - check-in
  /api/check-in:
    post:
      operationId: checkIn
      summary: ''
      description: チェックインする
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCheckInRequest'
      responses:
        '200':
          description: チェックインしたパーキングの情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingRoads'
      tags: *ref_0
  /api/check-in/get-user-here:
    post:
      operationId: getUserHere
      summary: ''
      description: パーキングにいるユーザを取得する
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetUserHereRequest'
      responses:
        '200':
          description: パーキングにいるユーザを取得する
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'
      tags: *ref_0
  /api/close/pa-road-list:
    get:
      operationId: pa-road-list
      summary: ''
      description: パーキング道路のリストを取得する
      parameters: []
      responses:
        '200':
          description: パーキング道路のリスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingRoads'
      tags: &ref_1
        - close
  /api/close/status-list:
    get:
      operationId: status-list
      summary: ''
      description: 閉鎖ステータスのリストを取得する
      parameters: []
      responses:
        '200':
          description: 閉鎖ステータスのリスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CloseStatuses'
      tags: *ref_1
  /api/close:
    post:
      operationId: post
      summary: ''
      description: 閉鎖状況を投稿する
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClosePostRequest'
      responses:
        '200':
          description: 投稿結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Closes'
      tags: *ref_1
  /api/close/status:
    get:
      operationId: status
      summary: ''
      description: 閉鎖状況を取得する
      parameters: []
      responses:
        '200':
          description: 閉鎖状況を取得する
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloseStatusLists'
      tags: *ref_1
  /api/contact:
    post:
      operationId: create
      summary: ''
      description: お問合せ
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '200':
          description: 問合せが成功したか
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contacts'
      tags:
        - contact
  /api/auth/signup:
    post:
      operationId: signup
      summary: ''
      description: サインアップする
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSignUpRequest'
      responses:
        '200':
          description: サインアップ結果
          content:
            application/json:
              schema:
                type: string
      tags: &ref_2
        - auth
  /api/auth/login:
    post:
      operationId: login
      summary: ''
      description: ログインする
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: JWTトークン
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtToken'
      tags: *ref_2
  /api/auth/active:
    post:
      operationId: active
      summary: ''
      description: サインイン後アドレス認証を行う
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveRequest'
      responses:
        '200':
          description: ユーザ情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtToken'
      tags: *ref_2
  /api/auth/me:
    get:
      operationId: me
      summary: ''
      description: ユーザ情報を取得する
      parameters: []
      responses:
        '200':
          description: ユーザ情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
      tags: *ref_2
info:
  title: Cats example
  description: The cats API description
  version: '1.0'
  contact: {}
tags:
  - name: cats
    description: ''
servers: []
components:
  schemas:
    GetCurrentParkingRequest:
      type: object
      properties:
        latitude:
          type: number
          description: 経度
          example: 35.46164868963681
        longitude:
          type: number
          description: 緯度
          example: 139.67996120452884
      required:
        - latitude
        - longitude
    CloseStatuses:
      type: object
      properties:
        id:
          type: number
          example: 1
        status:
          type: string
          example: close
        statusJpName:
          type: string
          example: 閉鎖
        colorCode:
          type: string
          example: '#FF00FF'
        group:
          type: string
          example: close
        close:
          type: array
          items:
            $ref: '#/components/schemas/Closes'
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
      required:
        - id
        - status
        - statusJpName
        - colorCode
        - group
        - close
        - createdAt
        - updatedAt
    Contacts:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: Primary key
        user:
          $ref: '#/components/schemas/Users'
        message:
          type: string
          example: 問い合わせ内容
          description: 問い合わせ内容
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
      required:
        - id
        - user
        - message
        - createdAt
        - updatedAt
    Users:
      type: object
      properties:
        id:
          type: number
          example: 1
        screenName:
          type: string
          example: screenname
        name:
          type: string
          example: name
        email:
          type: string
          example: example@example.com
        password:
          type: string
          example: password
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        checkIns:
          type: array
          items:
            type: string
        close:
          type: array
          items:
            $ref: '#/components/schemas/Closes'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contacts'
      required:
        - id
        - screenName
        - name
        - email
        - password
        - createdAt
        - updatedAt
        - checkIns
        - close
        - contacts
    Closes:
      type: object
      properties:
        id:
          type: int
          example: 1
        closeStatus:
          $ref: '#/components/schemas/CloseStatuses'
        parkingRoad:
          $ref: '#/components/schemas/ParkingRoads'
        user:
          $ref: '#/components/schemas/Users'
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
      required:
        - id
        - closeStatus
        - parkingRoad
        - user
        - createdAt
        - updatedAt
    Parkings:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 大黒PA
        latitude:
          type: number
          example: 35.46164868963681
        longitude:
          type: number
          example: 139.67996120452884
        radius:
          type: number
          example: 200
        parkingRoads:
          type: array
          items:
            $ref: '#/components/schemas/ParkingRoads'
        close:
          type: array
          items:
            $ref: '#/components/schemas/Closes'
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
      required:
        - id
        - name
        - latitude
        - longitude
        - radius
        - parkingRoads
        - close
        - createdAt
        - updatedAt
    CheckIns:
      type: object
      properties:
        id:
          type: int
          example: 1
        user:
          $ref: '#/components/schemas/Users'
        parkingRoad:
          $ref: '#/components/schemas/ParkingRoads'
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
      required:
        - id
        - user
        - parkingRoad
        - createdAt
        - updatedAt
    ParkingRoads:
      type: object
      properties:
        id:
          type: number
          example: 1
        parking:
          $ref: '#/components/schemas/Parkings'
        name:
          type: string
          example: 道路名
        createdAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        updatedAt:
          type: string
          example: '2024-10-01T00:00:00.000Z'
        checkIns:
          type: array
          items:
            $ref: '#/components/schemas/CheckIns'
        closes:
          type: array
          items:
            $ref: '#/components/schemas/Closes'
      required:
        - id
        - parking
        - name
        - createdAt
        - updatedAt
        - checkIns
        - closes
    PostCheckInRequest:
      type: object
      properties:
        latitude:
          type: number
          description: 経度
          example: 35.46164868963681
        longitude:
          type: number
          description: 緯度
          example: 139.67996120452884
        roadId:
          type: number
          description: 道路ID
          example: 1
      required:
        - latitude
        - longitude
        - roadId
    GetUserHereRequest:
      type: object
      properties:
        parkingRoadId:
          type: number
          description: パーキング道路ID
          example: 1
      required:
        - parkingRoadId
    ClosePostRequest:
      type: object
      properties:
        parkingRoadId:
          type: number
          description: 道路ID
          example: 1
        closeStatusId:
          type: number
          description: 閉鎖ステータスID
          example: 1
      required:
        - parkingRoadId
        - closeStatusId
    CloseStatusLists:
      type: object
      properties:
        list:
          type: object
          description: 閉鎖状況のリスト
          additionalProperties: true
      required:
        - list
    CreateContactRequest:
      type: object
      properties:
        userId:
          type: number
          description: ユーザID
          example: 1
        message:
          type: string
          description: 問い合わせ内容
          example: お問い合わせの内容をここに記入してください。
      required:
        - message
    AuthSignUpRequest:
      type: object
      properties:
        email:
          type: string
          description: メールアドレス
          example: hoge@example.com
        password:
          type: string
          description: パスワード
          example: fwe&3FD7Du2
        name:
          type: string
          description: ハンドルネーム
          example: name
        screenName:
          type: string
          description: スクリーンネーム
          example: screenname
      required:
        - email
        - password
        - name
        - screenName
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          description: メールアドレス
          example: hoge@example.com
        password:
          type: string
          description: パスワード
          example: fwe&3FD7Du2
      required:
        - email
        - password
    JwtToken:
      type: object
      properties:
        access_token:
          type: string
          description: JWTトークン
          example: eyJhbGciOiJIUzI1NiI...
      required:
        - access_token
    ActiveRequest:
      type: object
      properties:
        token:
          type: string
          description: アクティブトークン
          example: a59989d4-2bbe-4371-ad27-b0c3ecbfc513
      required:
        - token